Extract all loading-related code into new LoadingPlugin (loading.rs module)
